first pass at new framework
