+++ Examples/soc.js
@@ -8,10 +8,9 @@
 // Refer to included LICENSE file for usage rights and restrictions
 //=====================================================================
 
-var tjs = require('../TeslaJS');
-var fs = require('fs');
 require('colors');
 var program = require('commander');
+var framework = require('./sampleFramework.js');
 
 //
 //
@@ -24,33 +23,13 @@ program
   .parse(process.argv);
 
 //
-//
-//
-function login_cb(result) {
-    if (result.error) {
-        console.error("Login failed!".red);
-        console.warn(JSON.stringify(result.error));
-        return;
-    }
-
-    var options = { authToken: result.authToken, carIndex: program.index || 0 };
-    tjs.vehicles(options, function (err, vehicle) {
-        console.log("\nVehicle " + vehicle.vin + " ( '" + vehicle.display_name + "' ) is: " + vehicle.state.toUpperCase().bold.green);
-
-        if (vehicle.state.toUpperCase() == "OFFLINE") {
-            console.log("\nResult: " + "Unable to contact vehicle, exiting!".bold.red);
-            return;
-        }
-
-        options.vehicleID = vehicle.id_s;
-        sampleMain(options);
-    });
-}
+var sample = new framework.SampleFramework(program, sampleMain);
+sample.run();
 
 //
 //
 //
-function sampleMain(options) {
+function sampleMain(tjs, options) {
     tjs.chargeState(options, function (err, chargeState) {
 
         var str = chargeState.charge_port_door_open === true ? "OPEN" : "CLOSED";
@@ -90,38 +69,4 @@ function sampleMain(options) {
         console.log("Projected range: " + Math.round(chargeState.est_battery_range).toString().green + ' miles');
         console.log("Ideal range: " + Math.round(chargeState.ideal_battery_range).toString().green + ' miles');
     });
-}
-
-//
-// Sample starts here
-//
-var tokenFound = false;
-
-try {
-    tokenFound = fs.statSync('.token').isFile();
-} catch (e) {
-}
-
-if (program.uri) {
-    console.log("Setting portal URI to: " + program.uri);
-    tjs.setPortalBaseURI(program.uri);
-}
-
-if (tokenFound) {
-    var token = JSON.parse(fs.readFileSync('.token', 'utf8'));
-
-    if (!token) {
-        program.help();
-    }
-
-    login_cb({ error: false, authToken: token });
-} else {
-    var username = program.username || process.env.TESLAJS_USER;
-    var password = program.password || process.env.TESLAJS_PASS;
-
-    if (!username || !password) {
-        program.help();
-    }
-
-    tjs.login(username, password, login_cb);
 }
